requirements.txt
# Core dependencies

httpx>=0.27.0                
pydantic>=2.5.0              
pydantic-settings>=2.1.0     
python-dotenv>=1.0.0         
tenacity>=8.2.0              
tiktoken>=0.5.0              
fastapi>=0.109.0             
uvicorn>=0.27.0              

# Untuk PDF
PyMuPDF>=1.24.0              
pytesseract>=0.3.10          
Pillow>=10.0.0               
python-docx>=1.1.0           
python-multipart>=0.0.9